feat(ses): export Permits
This exports the entirety of `src/permits.js` as the `Permits` object via the `tools` export.

We are using this information in `lavamoat-tofu` for SES-compat detection, and we would like to keep it up-to-date with SES.

There are other ways we could accomplish this, but exporting it is the easiest way for `lavamoat-tofu` to consume it.