feat(compartment-mapper): expose loadCompartmentForArchive
This extracts a function, `loadCompartmentForArchive()`, from `digestLocation()` in `archive.js`, and exposes it.  This function returns archival-ready data structures, in addition to a record of the compartment renames which `makeArchiveCompartmentMap()` performs.

The intent is to leverage the module resolution performed by the function, which is reflected in its output.  A consumer can use the information therein to perform additional operations on the tree (e.g., parse ASTs).

# Conflicts:
#	packages/compartment-mapper/src/types.js